Support round ending

+++ backend/lib/game-persistence/src/main/kotlin/ru/nekoguys/game/persistence/competition/model/DbCompetitionRoundAnswer.kt
@@ -4,6 +4,6 @@ data class DbCompetitionRoundAnswer(
     var sessionId: Long,
     var roundNumber: Int,
     var teamId: Long,
-    var value: Long,
-    var income: Long?,
+    var value: Int,
+    var income: Double?,
 )
